codata.c: Fixed regression (nw)
git-svn-id: svn://messdev.no-ip.org/mess@14920 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/drivers/codata.c
@@ -17,7 +17,7 @@ public:
 		: driver_device(mconfig, type, tag) ,
 		m_p_base(*this, "p_base"){ }
 
-	required_shared_ptr<UINT8> m_p_base;
+	required_shared_ptr<UINT16> m_p_base;
 };
